Tear down detection timeout in test (from Long.MAX_VALUE to Integer.MAX_VALUE)

+++ modules/core/src/test/java/org/apache/ignite/internal/GridAffinityMappedTest.java
@@ -59,7 +59,7 @@ public class GridAffinityMappedTest extends GridCommonAbstractTest {
 
         cfg.setDiscoverySpi(disco);
 
-        cfg.setFailureDetectionTimeout(Long.MAX_VALUE);
+        cfg.setFailureDetectionTimeout(Integer.MAX_VALUE);
 
         if (igniteInstanceName.endsWith("1"))
             cfg.setCacheConfiguration(); // Empty cache configuration.
